Move test-only schema stuff to a separate crate, which is only included as a dev-dependency
