+++ cyclonedx-bom/Cargo.toml
@@ -17,7 +17,6 @@ rust-version.workspace = true
 base64 = "0.21.2"
 fluent-uri = "0.1.4"
 indexmap = "2.2.2"
-jsonschema = "0.17.1"
 once_cell = "1.18.0"
 ordered-float = { version = "4.2.0", default-features = false }
 packageurl = "0.3.0"
@@ -35,3 +34,4 @@ strum = { version = "0.26.2", features = ["derive"] }
 [dev-dependencies]
 insta = { version = "1.33.0", features = ["glob", "json"] }
 pretty_assertions = "1.4.0"
+test-utils = {path = "test-utils"}
