+++ cyclonedx-bom/src/lib.rs
@@ -126,7 +126,6 @@ pub mod errors;
 pub mod external_models;
 pub mod models;
 pub mod prelude;
-pub mod schema;
 pub mod validation;
 
 mod specs;
