+++ cyclonedx-bom/test-utils/schema/jsf-0.82.schema.json
