+++ cyclonedx-bom/test-utils/schema/spdx.schema.json
